@@ -23,7 +23,7 @@
         "source": "https://github.com/danielspk/TORNADO"
     },
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4.0"
     },
     "autoload": {
         "psr-4": {
